Organization, documentation tweaks
