+++ Sources/NSMetadataToolbox/Documentation.docc/Collections/Queries.md
@@ -2,6 +2,14 @@
 
 ## Topics
 
+### Notification Names
+
+- ``Foundation/NSNotification/Name/NSMetadataQuery/didStartGathering``
+- ``Foundation/NSNotification/Name/NSMetadataQuery/didFinishGathering``
+- ``Foundation/NSNotification/Name/NSMetadataQuery/gatheringProgress``
+- ``Foundation/NSNotification/Name/NSMetadataQuery/didUpdate``
+
+
 ### Concurrency
 
 - ``Foundation/NSMetadataQuery/gatherResults(isolation:notificationCenter:)``
