+++ Sources/NSMetadataToolbox/Documentation.docc/Documentation.md
@@ -9,3 +9,8 @@ Extensions and conveniences for Apple's metadata APIs.
 - <doc:AttributeKeys>
 - <doc:AttributeModifiers>
 - <doc:Queries>
+
+
+### Errors
+
+- ``NSMetadataError``
